middleware 수정
- middleware 경로 수정
  - root -> /src/middleware.ts
- 코드 리팩토링
  - 경로 상수 분리
  - 조건 분기 단순화
    - authorized 콜백 -> 모든 요청 허용. 대신 미들웨어 함수에서 직접 인증 로직 처리
  - 무한 리다이렉트 루프 방지
